Send out different email reminder if registered Mon-Wed
As they won't have received the initial welcome email when they
registered we send out a slightly different email to people who
registered in the preceding Monday - Wednesday.

Also update code to not send out to people who registered on the day the
emails are being sent as they will have got an email as part of the
registration process.